refactor: extract early scan-flow steps + shared header/button (#80, pt 1)
First step decomposing ScanEntryFlow (1171 → 886). Extracts the smaller,
self-contained steps and the shared navigation controls; the larger review
form (post-OCR) and manual entry come in pt 2.

New sibling files:
  ScanFlowControls.swift  — ScanFlowPushHeader, ScanFlowContinueButton
                            (presentational, reusable)
  ScanTypeStep.swift      — step 1: pick scan type
  InputMethodStep.swift   — step 2: scan vs manual entry
  CameraStep.swift        — step 3: document scanner + processing overlay

ScanEntryFlow's body now dispatches the early steps to the new structs and
passes a single onClose closure for the type-step's back-to-dismiss action.
Verbatim move — no behavior change.

Verified: make lint (0, --strict), make build, make test (272 logic + 13 UI
incl. ScanEntryFlowUITests, all green).
