+++ Baseline/Views/Body/ScanEntryFlow.swift
@@ -41,11 +41,11 @@ struct ScanEntryFlow: View {
                     // Swift's type decoder on ARM devices.
                     switch vm.currentStep {
                     case .selectType:
-                        AnyView(scanTypeStep(vm: vm))
+                        AnyView(ScanTypeStep(vm: vm, onClose: { dismiss() }))
                     case .selectMethod:
-                        AnyView(inputMethodStep(vm: vm))
+                        AnyView(InputMethodStep(vm: vm))
                     case .camera:
-                        AnyView(cameraStep(vm: vm))
+                        AnyView(CameraStep(vm: vm))
                     case .review:
                         AnyView(reviewFormStep(vm: vm))
                     case .manualEntry:
@@ -74,235 +74,6 @@ struct ScanEntryFlow: View {
             }
         } message: {
             Text(resolvedVM?.errorMessage ?? "Something went wrong while saving. Try again.")
-        }
-    }
-
-    // MARK: - Step 1: Scan Type Selection
-
-    private func scanTypeStep(vm: ScanEntryViewModel) -> some View {
-        VStack(spacing: 0) {
-            pushHeader(title: "New Scan", subtitle: "Step 1 of 2") {
-                dismiss()
-            }
-
-            ScrollView {
-                VStack(alignment: .leading, spacing: 0) {
-                    // Intro text
-                    VStack(alignment: .leading, spacing: 6) {
-                        Text("What type of scan?")
-                            .font(.system(size: 22, weight: .bold))
-                            .tracking(-0.4)
-                            .foregroundStyle(CadreColors.textPrimary)
-                        Text("Each machine records a different set of metrics.")
-                            .font(.system(size: 12, weight: .medium))
-                            .foregroundStyle(CadreColors.textTertiary)
-                    }
-                    .padding(.horizontal, CadreSpacing.sheetHorizontal)
-                    .padding(.top, 20)
-                    .padding(.bottom, 12)
-
-                    // InBody 570 card (pre-selected)
-                    scanTypeCard(
-                        icon: "570",
-                        name: "InBody 570",
-                        description: "Body comp \u{00B7} segmental lean \u{00B7} water analysis",
-                        isSelected: vm.selectedType == .inBody
-                    )
-                    .padding(.horizontal, CadreSpacing.sheetHorizontal)
-
-                    // "More coming soon" note
-                    comingSoonNote
-                        .padding(.horizontal, CadreSpacing.sheetHorizontal)
-                        .padding(.top, 14)
-                }
-            }
-
-            // Continue button at bottom
-            continueButton(label: "Continue") {
-                vm.selectType(.inBody)
-            }
-        }
-    }
-
-    private func scanTypeCard(icon: String, name: String, description: String, isSelected: Bool) -> some View {
-        HStack(spacing: 12) {
-            // Icon
-            Text(icon)
-                .font(.system(size: 10, weight: .bold, design: .default))
-                .tracking(-0.2)
-                .foregroundStyle(CadreColors.accent)
-                .frame(width: 36, height: 36)
-                .background(isSelected ? CadreColors.accent.opacity(0.2) : CadreColors.cardElevated)
-                .clipShape(RoundedRectangle(cornerRadius: 10))
-
-            VStack(alignment: .leading, spacing: 3) {
-                Text(name)
-                    .font(.system(size: 14, weight: .semibold))
-                    .tracking(-0.1)
-                    .foregroundStyle(CadreColors.textPrimary)
-                Text(description)
-                    .font(.system(size: 10, weight: .medium))
-                    .foregroundStyle(CadreColors.textTertiary)
-            }
-
-            Spacer(minLength: 0)
-        }
-        .padding(14)
-        .background(isSelected ? CadreColors.cardElevated : CadreColors.card)
-        .overlay(
-            RoundedRectangle(cornerRadius: 14)
-                .stroke(isSelected ? CadreColors.accent : CadreColors.divider, lineWidth: 1)
-        )
-        .clipShape(RoundedRectangle(cornerRadius: 14))
-    }
-
-    private var comingSoonNote: some View {
-        HStack(spacing: 8) {
-            Image(systemName: "info.circle")
-                .font(.system(size: 13, weight: .medium))
-                .foregroundStyle(CadreColors.textTertiary)
-            Text("More scan types coming soon")
-                .font(.system(size: 11, weight: .medium))
-                .foregroundStyle(CadreColors.textTertiary)
-        }
-        .padding(.horizontal, 14)
-        .padding(.vertical, 10)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(CadreColors.card)
-        .overlay(
-            RoundedRectangle(cornerRadius: 10)
-                .stroke(CadreColors.divider, lineWidth: 0.5)
-        )
-        .clipShape(RoundedRectangle(cornerRadius: 10))
-    }
-
-    // MARK: - Step 2: Input Method
-
-    private let multiPhotoTip = MultiPhotoTip()
-
-    private func inputMethodStep(vm: ScanEntryViewModel) -> some View {
-        VStack(spacing: 0) {
-            pushHeader(title: "New Scan", subtitle: "Step 2 of 2") {
-                vm.goBack()
-            }
-
-            VStack(alignment: .leading, spacing: 0) {
-                Text("How do you want to enter it?")
-                    .font(.system(size: 22, weight: .bold))
-                    .tracking(-0.4)
-                    .foregroundStyle(CadreColors.textPrimary)
-                    .padding(.horizontal, CadreSpacing.sheetHorizontal)
-                    .padding(.top, 20)
-                    .padding(.bottom, 8)
-
-                TipView(multiPhotoTip)
-                    .padding(.horizontal, CadreSpacing.sheetHorizontal)
-                    .padding(.bottom, 4)
-
-                VStack(spacing: 12) {
-                    methodCard(
-                        icon: "camera",
-                        title: "Scan printout",
-                        description: "Auto-reads values from your InBody printout.",
-                        hint: "Multiple photos improve accuracy"
-                    ) {
-                        vm.selectMethod(camera: true)
-                    }
-
-                    methodCard(
-                        icon: "square.and.pencil",
-                        title: "Enter manually",
-                        description: "Type the values from your printout yourself."
-                    ) {
-                        vm.selectMethod(camera: false)
-                    }
-                    .accessibilityIdentifier(A11yID.ScanEntry.manualEntryButton)
-                }
-                .padding(.horizontal, CadreSpacing.sheetHorizontal)
-                .padding(.top, 20)
-            }
-
-            Spacer()
-        }
-    }
-
-    private func methodCard(
-        icon: String, title: String, description: String, hint: String? = nil, action: @escaping () -> Void
-    ) -> some View {
-        Button(action: action) {
-            HStack(spacing: 14) {
-                Image(systemName: icon)
-                    .font(.system(size: 22, weight: .regular))
-                    .foregroundStyle(CadreColors.accent)
-                    .frame(width: 44, height: 44)
-                    .background(CadreColors.cardElevated)
-                    .clipShape(RoundedRectangle(cornerRadius: 12))
-
-                VStack(alignment: .leading, spacing: 4) {
-                    Text(title)
-                        .font(.system(size: 15, weight: .bold))
-                        .tracking(-0.2)
-                        .foregroundStyle(CadreColors.textPrimary)
-                    Text(description)
-                        .font(.system(size: 12, weight: .medium))
-                        .foregroundStyle(CadreColors.textTertiary)
-                        .lineSpacing(2)
-                    if let hint {
-                        HStack(spacing: 4) {
-                            Image(systemName: "lightbulb.min")
-                                .font(.system(size: 10, weight: .medium))
-                            Text(hint)
-                                .font(.system(size: 11, weight: .medium))
-                        }
-                        .foregroundStyle(CadreColors.accent.opacity(0.7))
-                        .padding(.top, 2)
-                    }
-                }
-
-                Spacer(minLength: 0)
-
-                Image(systemName: "chevron.right")
-                    .font(.system(size: 12, weight: .semibold))
-                    .foregroundStyle(CadreColors.textTertiary)
-            }
-            .padding(.horizontal, 18)
-            .padding(.vertical, 20)
-            .background(CadreColors.card)
-            .overlay(
-                RoundedRectangle(cornerRadius: 16)
-                    .stroke(CadreColors.divider, lineWidth: 1)
-            )
-            .clipShape(RoundedRectangle(cornerRadius: 16))
-        }
-        .buttonStyle(.plain)
-    }
-
-    // MARK: - Step 3: Camera
-
-    private func cameraStep(vm: ScanEntryViewModel) -> some View {
-        ZStack {
-            DocumentScannerView(
-                onScan: { scan in
-                    Task {
-                        await vm.processScan(scan)
-                    }
-                },
-                onCancel: {
-                    vm.goBack()
-                }
-            )
-            .ignoresSafeArea()
-
-            if vm.isProcessing {
-                Color.black.opacity(0.6)
-                    .ignoresSafeArea()
-                VStack(spacing: 16) {
-                    ProcessingIndicator(size: 56, lineWidth: 5)
-                    Text("Reading scan...")
-                        .font(.system(size: 14, weight: .semibold))
-                        .foregroundStyle(CadreColors.textPrimary)
-                }
-            }
         }
     }
 
@@ -1105,62 +876,6 @@ struct ScanEntryFlow: View {
         } catch {
             vm.errorMessage = error.localizedDescription
         }
-    }
-
-    // MARK: - Push Header
-
-    private func pushHeader(title: String, subtitle: String, backAction: @escaping () -> Void) -> some View {
-        HStack(spacing: 12) {
-            Button(action: backAction) {
-                Image(systemName: "chevron.left")
-                    .font(.system(size: 14, weight: .semibold))
-                    .foregroundStyle(CadreColors.textPrimary)
-                    .frame(width: 30, height: 30)
-                    .background(CadreColors.cardElevated)
-                    .clipShape(Circle())
-            }
-            .buttonStyle(.plain)
-
-            Spacer()
-
-            VStack(spacing: 2) {
-                Text(title)
-                    .font(.system(size: 17, weight: .bold))
-                    .tracking(-0.2)
-                    .foregroundStyle(CadreColors.textPrimary)
-                Text(subtitle)
-                    .font(.system(size: 10, weight: .medium))
-                    .foregroundStyle(CadreColors.textTertiary)
-            }
-
-            Spacer()
-
-            // Invisible spacer for centering
-            Color.clear
-                .frame(width: 30, height: 30)
-        }
-        .padding(.horizontal, CadreSpacing.sheetHorizontal)
-        .padding(.top, 14)
-        .padding(.bottom, 8)
-    }
-
-    // MARK: - Continue Button
-
-    private func continueButton(label: String, action: @escaping () -> Void) -> some View {
-        Button(action: action) {
-            Text(label)
-                .font(.system(size: 15, weight: .semibold))
-                .tracking(0.3)
-                .foregroundStyle(.white)
-                .frame(maxWidth: .infinity)
-                .padding(.vertical, 15)
-                .background(CadreColors.accent)
-                .clipShape(RoundedRectangle(cornerRadius: 14))
-        }
-        .accessibilityIdentifier(A11yID.ScanEntry.continueButton)
-        .padding(.horizontal, CadreSpacing.sheetHorizontal)
-        .padding(.top, 20)
-        .padding(.bottom, 16)
     }
 }
